修复 Migration 7→8: 严格匹配 Room 期望的 schema 格式
- tags 表 CREATE TABLE 列顺序与 Room schema 完全一致
- exercise_tags 表去除不必要的额外索引
- 外键顺序与实体定义一致
- 更新预填充 kenko.db 用户版本为 6
- build.gradle 配置缓存清理

+++ app/src/main/kotlin/com/looker/kenko/data/repository/local/LocalExerciseRepo.kt
@@ -49,7 +49,7 @@ class LocalExerciseRepo @Inject constructor(
 
     override suspend fun upsert(exercise: Exercise) {
         val entityId = dao.upsert(exercise.toEntity())
-        val id = exercise.id ?: entityId
+        val id = exercise.id ?: entityId.toInt()
         if (id != 0) {
             tagDao.replaceExerciseTags(id, exercise.tags.map { it.id })
         }
